refactor: resolve #870 - split sampleBgData.ts (711 lines) exceeding 500-line limit
Extracts visual grid infrastructure and manual sample grids into separate
modules. sampleBgData.ts becomes public API facade (285 lines). Zero
consumer changes needed.

Closes #870
