+++ src/core/samples/sampleBgData.ts
@@ -1,37 +1,21 @@
-/* eslint-disable max-len */
-/* eslint-disable @typescript-eslint/naming-convention */
+ 
 /**
  * Sample BG Data for programs that use the VIEW command
  *
- * These are pre-designed BG graphics that pair with sample programs.
- * Each BG is a 28x21 grid of cells with character codes and color patterns.
+ * Public API for sample BG data. Provides pre-designed BG graphics that
+ * pair with sample programs. Each BG is a 28x21 grid of cells with
+ * character codes and color patterns.
  */
 
-import { BG_GRID, DEFAULT_BG_CHAR_CODE } from '@/features/bg-editor/constants'
-import type { BgCell, BgGridData } from '@/features/bg-editor/types'
+import { BG_GRID } from '@/features/bg-editor/constants'
+import type { BgGridData } from '@/features/bg-editor/types'
 
-/** Create an empty 28x21 grid */
-function createEmptyGrid(): BgGridData {
-  const grid: BgGridData = []
-  for (let y = 0; y < BG_GRID.ROWS; y++) {
-    const row: BgCell[] = []
-    for (let x = 0; x < BG_GRID.COLS; x++) {
-      row.push({ charCode: DEFAULT_BG_CHAR_CODE, colorPattern: 0 })
-    }
-    grid.push(row)
-  }
-  return grid
-}
+import { MANUAL_SAMPLE_BG_GRIDS } from './sampleBgManualGrids'
+import { createEmptyGrid, setCell } from './sampleBgVisualGrid'
 
-/** Helper to set a cell in the grid */
-function setCell(grid: BgGridData, x: number, y: number, charCode: number, colorPattern: 0 | 1 | 2 | 3 = 0): void {
-  if (x >= 0 && x < BG_GRID.COLS && y >= 0 && y < BG_GRID.ROWS) {
-    const row = grid[y]
-    if (row) {
-      row[x] = { charCode, colorPattern }
-    }
-  }
-}
+// ============================================================================
+// IDE Demo Grid Creators
+// ============================================================================
 
 /**
  * Demo BG with a simple border and centered text area
@@ -233,395 +217,6 @@ function createTestPatternGrid(): BgGridData {
   return grid
 }
 
-// ============================================================================
-// F-BASIC Reference Manual Sample Games BG Data (pages 94-101)
-// ============================================================================
-
-/** Empty cell marker for visual grids (3 underscores for visual alignment) */
-const ___ = '___'
-
-/** Letter to base CHR$ value mapping (from manual page 113) */
-const LETTER_BASE: Record<string, number> = {
-  A: 0, B: 8, C: 16, D: 24,
-  E: 184, F: 192, G: 200, H: 208,
-  I: 216, J: 224, K: 232, L: 240, M: 248,
-}
-
-// BG code constants for visual grids (from manual page 113)
-// Format: {letter}{digit}{color} where color is 0-3
-// K codes (232-239)
-const K72 = 'K72', K62 = 'K62', K52 = 'K52', K42 = 'K42'
-const K32 = 'K32', K22 = 'K22', K12 = 'K12', K02 = 'K02'
-const K50 = 'K50'
-// L codes (240-247)
-const L02 = 'L02', L12 = 'L12', L22 = 'L22'
-const L40 = 'L40'
-// J codes (224-231)
-const J00 = 'J00', J10 = 'J10', J20 = 'J20', J30 = 'J30'
-const J02 = 'J02', J12 = 'J12', J22 = 'J22', J32 = 'J32'
-const J50 = 'J50', J33 = 'J33', J03 = 'J03', J13 = 'J13', J23 = 'J23', J70 = 'J70'
-// I codes (216-223)
-const I60 = 'I60', I70 = 'I70', I62 = 'I62', I72 = 'I72'
-const I63 = 'I63', I73 = 'I73'
-// G codes (200-207)
-const G00 = 'G00', G10 = 'G10', G20 = 'G20', G30 = 'G30', G40 = 'G40'
-const G42 = 'G42', G52 = 'G52', G62 = 'G62'
-// D codes (24-31)
-const D41 = 'D41', D50 = 'D50', D52 = 'D52'
-// F codes (192-199)
-const F32 = 'F32', F41 = 'F41', F61 = 'F61', F72 = 'F72'
-// M codes (248-255)
-const M11 = 'M11', M31 = 'M31', M70 = 'M70', M71 = 'M71', M72 = 'M72'
-
-// ============================================================================
-// ASCII Character Constants for Visual Grids (3-char padded for alignment)
-// ============================================================================
-const _1_ = '1  ', _2_ = '2  ', _3_ = '3  ', _4_ = '4  ', _5_ = '5  '
-const _6_ = '6  ', _7_ = '7  ', _8_ = '8  '
-const _A_ = 'A  ', _B_ = 'B  ', _C_ = 'C  ', _D_ = 'D  ', _E_ = 'E  '
-const _F_ = 'F  ', _G_ = 'G  ', _H_ = 'H  ', _I_ = 'I  ', _J_ = 'J  '
-const _K_ = 'K  ', _L_ = 'L  ', _M_ = 'M  ', _N_ = 'N  ', _O_ = 'O  '
-const _P_ = 'P  ', _R_ = 'R  ', _S_ = 'S  ', _T_ = 'T  ', _U_ = 'U  '
-const _Y_ = 'Y  '
-
-/**
- * Convert a 3-char BG code from manual (e.g., "K72", "L02") to [charCode, colorPattern]
- */
-function parseBgCode(code: string): [number, 0 | 1 | 2 | 3] {
-  if (code.length !== 3) {
-    throw new Error(`Invalid BG code: ${code}`)
-  }
-
-  const letter = code[0]!
-  const digit = parseInt(code[1]!, 10)
-  const colorChar = code[2]!
-  const colorPattern = parseInt(colorChar, 10) as 0 | 1 | 2 | 3
-
-  const base = LETTER_BASE[letter]
-  if (base === undefined) {
-    throw new Error(`Unknown letter in BG code: ${code}`)
-  }
-
-  const charCode = base + digit
-  return [charCode, colorPattern]
-}
-
-/**
- * Create a cell from ASCII character
- */
-function asciiCell(char: string, color: 0 | 1 | 2 | 3 = 0): [number, 0 | 1 | 2 | 3] {
-  return [char.charCodeAt(0), color]
-}
-
-/**
- * Convert a visual 2D string grid to BgGridData
- *
- * Cell format:
- * - "___" (3 underscores) = empty cell
- * - "K72" = BG code (letter + digit + color pattern)
- * - "A  " = ASCII character (letter + 2 spaces)
- *
- * @param visualGrid - 2D array of 3-char cell strings
- * @returns BgGridData ready for use
- */
-function visualGridToBgGridData(visualGrid: string[][]): BgGridData {
-  const result = createEmptyGrid()
-
-  for (let y = 0; y < visualGrid.length && y < BG_GRID.ROWS; y++) {
-    const row = visualGrid[y]
-    if (!row) continue
-
-    for (let x = 0; x < row.length && x < BG_GRID.COLS; x++) {
-      const cell = row[x]
-      if (!cell || cell === ___) continue
-
-      const trimmed = cell.trim()
-      if (trimmed.length === 0) continue
-
-      // Check if it's a 3-char BG code (letter + digit + color) or ASCII text
-      if (/^[A-M][0-7][0-3]$/.test(trimmed)) {
-        // BG code like K72, L02, etc.
-        const [charCode, colorPattern] = parseBgCode(trimmed)
-        setCell(result, x, y, charCode, colorPattern)
-      } else {
-        // ASCII text - use first character
-        const [charCode, colorPattern] = asciiCell(trimmed[0]!, 0)
-        setCell(result, x, y, charCode, colorPattern)
-      }
-    }
-  }
-
-  return result
-}
-
-/**
- * KNIGHT game BG (page 94)
- * 8x8 chessboard with labels A-H and 8-1
- */
-function createKnightGrid(): BgGridData {
-  const grid: string[][] = [
-    //       0    1    2    3    4    5    6    7    8    9   10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
-    /*  0 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  1 */ [___, _8_, ___, _7_, ___, _6_, ___, _5_, ___, _4_, ___, _3_, ___, _2_, ___, _1_, ___, ___, ___, I60, J30, J30, J30, J30, J30, J30, I70, ___],
-    /*  2 */ [K72, K52, K22, K52, K22, K52, K22, K52, K22, K52, K22, K52, K22, K52, K22, K52, L02, ___, ___, J20, _K_, _N_, _I_, _G_, _H_, _T_, J20, ___],
-    /*  3 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _A_, ___, J00, J30, J30, J30, J30, J30, J30, J10, ___],
-    /*  4 */ [K42, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  5 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _B_, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  6 */ [K42, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  7 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _C_, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  8 */ [K42, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  9 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _D_, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 10 */ [K42, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 11 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _E_, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 12 */ [K42, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 13 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _F_, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 14 */ [K42, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 15 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _G_, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 16 */ [K42, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 17 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _H_, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 18 */ [L12, K52, K12, K52, K12, K52, K12, K52, K12, K52, K12, K52, K12, K52, K12, K52, L22, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 19 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 20 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-  ]
-
-  return visualGridToBgGridData(grid)
-}
-
-/**
- * SUPER MEMORY game BG (page 95)
- * Four L-shaped corner panels pointing toward center
- */
-function createSuperMemoryGrid(): BgGridData {
-  const grid: string[][] = [
-    //         0    1    2    3    4    5    6    7    8    9    10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
-    /*  0 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  1 */ [___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, I60, J30, J30, J30, J30, J30, I70],
-    /*  2 */ [___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, J20, _S_, _U_, _P_, _E_, _R_, J50],
-    /*  3 */ [___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, J20, _M_, _E_, _M_, _O_, _R_, _Y_],
-    /*  4 */ [___, K52, K52, K52, L22, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, L12, K52, K52, K52, ___, J00, J30, J30, J30, J30, J30, J70],
-    /*  5 */ [___, K52, K52, K52, K52, L22, K62, ___, ___, ___, ___, ___, ___, ___, K62, L12, K52, K52, K52, K52, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  6 */ [___, K52, K52, K52, K52, K52, L22, ___, ___, ___, ___, ___, ___, ___, L12, K52, K52, K52, K52, K52, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  7 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  8 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  9 */ [___, ___, ___, ___, ___, ___, ___, ___, K72, K52, K52, K52, L02, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 10 */ [___, ___, ___, ___, ___, ___, ___, ___, K62, ___, ___, ___, K62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 11 */ [___, ___, ___, ___, ___, ___, ___, ___, L12, K52, K52, K52, L22, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 12 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 13 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 14 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 15 */ [___, K52, K52, K52, K52, K52, L02, ___, ___, ___, ___, ___, ___, ___, K72, K52, K52, K52, K52, K52, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 16 */ [___, K52, K52, K52, K52, L02, K62, ___, ___, ___, ___, ___, ___, ___, K62, K72, K52, K52, K52, K52, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 17 */ [___, K52, K52, K52, L02, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, K72, K52, K52, K52, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 18 */ [___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 19 */ [___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 20 */ [___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-  ]
-
-  return visualGridToBgGridData(grid)
-}
-
-/**
- * UFO game BG (page 96)
- * Starry sky with mountain terrain at bottom
- */
-function createUfoGrid(): BgGridData {
-  const grid: string[][] = [
-    //         0    1    2    3    4    5    6    7    8    9    10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
-    /*  0 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  1 */ [___, ___, ___, M11, M31, ___, ___, ___, ___, G62, ___, ___, G52, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  2 */ [___, F41, D41, D41, D41, D41, F61, ___, ___, ___, ___, ___, ___, ___, ___, G62, ___, ___, ___, G62, ___, ___, ___, ___, G52, ___, ___, ___],
-    /*  3 */ [___, ___, K12, K52, K52, K12, ___, ___, ___, ___, ___, ___, G62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  4 */ [___, G52, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G62, ___, ___, ___, ___, ___, ___, G62],
-    /*  5 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  6 */ [___, ___, ___, G62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  7 */ [___, ___, ___, ___, ___, ___, G62, ___, ___, ___, G62, ___, ___, ___, ___, ___, ___, ___, G52, ___, ___, ___, ___, G62, ___, ___, ___, ___],
-    /*  8 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G62, ___, ___],
-    /*  9 */ [___, G62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 10 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G52, G62, ___, ___, ___, ___, ___, G62, ___, G52, ___, ___, ___],
-    /* 11 */ [___, ___, ___, ___, ___, ___, ___, G52, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G62, ___],
-    /* 12 */ [___, ___, ___, G62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G62, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 13 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 14 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G00, G10, ___, ___, ___, ___],
-    /* 15 */ [___, ___, ___, ___, ___, ___, ___, ___, G00, G10, ___, ___, ___, ___, ___, ___, ___, ___, ___, G00, G10, G00, G20, G30, G10, ___, ___, ___],
-    /* 16 */ [___, ___, ___, ___, ___, ___, ___, G00, G20, G30, G10, ___, ___, ___, ___, ___, ___, ___, G00, G20, G20, G20, G20, G30, G30, G10, ___, ___],
-    /* 17 */ [G40, M70, M70, G40, G40, M70, G20, G20, G20, G40, G30, G40, G40, M70, G40, M70, G40, G40, G40, G40, G20, G20, G40, G40, G30, G30, G40, G40],
-    /* 18 */ [M70, G40, M70, M70, G40, G40, G40, G40, M70, M70, G40, G40, M70, G40, M70, M70, M70, G40, M70, M70, G40, G40, M70, G40, G40, G40, M70, M70],
-    /* 19 */ [F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32],
-    /* 20 */ [M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72],
-  ]
-
-  return visualGridToBgGridData(grid)
-}
-
-/**
- * ROUTE 66 game BG (page 97)
- * Road with edges - all rows same pattern
- */
-function createRoute66Grid(): BgGridData {
-  const grid: string[][] = [
-    //         0    1    2    3    4    5    6    7    8    9    10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
-    /*  0 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /*  1 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /*  2 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /*  3 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /*  4 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /*  5 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /*  6 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /*  7 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /*  8 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /*  9 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 10 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 11 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 12 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 13 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 14 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 15 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 16 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 17 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 18 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 19 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-    /* 20 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
-  ]
-
-  return visualGridToBgGridData(grid)
-}
-
-/**
- * TYPE MASTER game BG (page 98)
- * Text boxes and frames for typing display
- */
-function createTypeMasterGrid(): BgGridData {
-  const grid: string[][] = [
-    //         0    1    2    3    4    5    6    7    8    9    10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
-    /*  0 */ [___, K72, K52, K52, K52, K52, L02, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  1 */ [___, K62, _T_, _Y_, _P_, _E_, K62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  2 */ [___, L12, K52, K52, K52, K52, L22, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  3 */ [___, I60, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, I70, ___, ___, ___, ___],
-    /*  4 */ [___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J20, ___, ___, ___, ___],
-    /*  5 */ [___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J20, ___, ___, ___, ___],
-    /*  6 */ [___, J00, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J10, ___, ___, ___, ___],
-    /*  7 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  8 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /*  9 */ [___, I60, J30, J30, J30, J30, J30, J30, J30, J30, J30, I70, I62, J32, I72, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 10 */ [___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, J20, J22, ___, J22, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 11 */ [___, J00, J30, J30, J30, J30, J30, J30, J30, J30, J30, J10, J02, J32, J12, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 12 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 13 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 14 */ [___, I63, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, I73, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 15 */ [___, J23, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J23, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 16 */ [___, J23, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J23, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 17 */ [___, J03, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J13, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 18 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 19 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 20 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-  ]
-
-  return visualGridToBgGridData(grid)
-}
-
-/**
- * TURTLE game BG (page 99)
- * Racing track with lanes and finish line
- */
-function createTurtleGrid(): BgGridData {
-  const grid: string[][] = [
-    //         0    1    2    3    4    5    6    7    8    9    10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
-    /*  0 */ [D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50],
-    /*  1 */ [D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52],
-    /*  2 */ [G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42],
-    /*  3 */ [G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52],
-    /*  4 */ [G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52],
-    /*  5 */ [K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K52, K52, K52, K22],
-    /*  6 */ [___, ___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
-    /*  7 */ [_1_, ___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
-    /*  8 */ [K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K52, K52, K52, K02],
-    /*  9 */ [___, ___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
-    /* 10 */ [_2_, ___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
-    /* 11 */ [K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K52, K52, K52, K02],
-    /* 12 */ [___, ___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
-    /* 13 */ [ _3_,___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
-    /* 14 */ [K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K52, K52, K52, K02],
-    /* 15 */ [___, ___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
-    /* 16 */ [ _4_,___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
-    /* 17 */ [K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K52, K52, K52, K02],
-    /* 18 */ [___, ___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
-    /* 19 */ [ _5_,___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
-    /* 20 */ [K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K52, K52, K52, K12],
-  ]
-
-  return visualGridToBgGridData(grid)
-}
-
-/**
- * CARD game BG (page 100)
- * 6x6 card grid frame
- */
-function createCardGrid(): BgGridData {
-  const grid: string[][] = [
-    //         0    1    2    3    4    5    6    7    8    9    10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
-    /*  0 */ [K72, K52, K52, L02, K72, K52, K52, L02, K72, K52, K52, L02, K72, K52, K52, L02, K72, K52, K52, L02, K72, K52, K52, L02, ___, ___, ___, ___],
-    /*  1 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /*  2 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /*  3 */ [K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, ___, ___, ___, ___],
-    /*  4 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /*  5 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /*  6 */ [K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, ___, ___, ___, ___],
-    /*  7 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /*  8 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /*  9 */ [K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, F72, ___, ___, ___],
-    /* 10 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /* 11 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /* 12 */ [K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, F72, ___, ___, ___],
-    /* 13 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /* 14 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /* 15 */ [K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, ___, ___, ___, ___],
-    /* 16 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /* 17 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
-    /* 18 */ [L12, K52, K52, L22, L12, K52, K52, L22, L12, K52, K52, L22, L12, K52, K52, L22, L12, K52, K52, L22, L12, K52, K52, L22, ___, ___, ___, ___],
-    /* 19 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-    /* 20 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
-  ]
-
-  return visualGridToBgGridData(grid)
-}
-
-/**
- * SCR$ Sample game BG (page 101)
- * Maze with flags (F72) and brick walls (F32)
- */
-function createScrSampleGrid(): BgGridData {
-  const grid: string[][] = [
-    //         0    1    2    3    4    5    6    7    8    9    10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
-    /*  0 */ [F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32],
-    /*  1 */ [F32, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32],
-    /*  2 */ [F32, ___, ___, F72, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, F72, ___, ___, ___, F72, ___, ___, ___, F72, ___, ___, ___, ___, F32],
-    /*  3 */ [F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32],
-    /*  4 */ [F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, F32, F32, F32, F32, F32, F32, F32, ___, ___, ___, ___, F32],
-    /*  5 */ [F32, ___, ___, ___, ___, ___, ___, ___, F72, ___, ___, ___, ___, ___, F32, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, F32],
-    /*  6 */ [F32, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, F72, ___, ___, F32, F32, ___, F72, ___, F32],
-    /*  7 */ [F32, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32],
-    /*  8 */ [F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32],
-    /*  9 */ [F32, ___, ___, ___, F72, ___, ___, ___, ___, ___, ___, ___, ___, F72, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, F32],
-    /* 10 */ [F32, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, F72, ___, ___, ___, ___, F32],
-    /* 11 */ [F32, F32, F32, F32, F32, F32, F32, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32],
-    /* 12 */ [F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32],
-    /* 13 */ [F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, F72, ___, ___, ___, ___, ___, ___, F72, ___, F32],
-    /* 14 */ [F32, ___, F72, ___, ___, ___, ___, ___, ___, ___, F72, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32],
-    /* 15 */ [F32, ___, ___, ___, ___, F32, F32, F32, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, F32],
-    /* 16 */ [F32, ___, ___, ___, ___, F32, F32, F32, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, F32],
-    /* 17 */ [F32, ___, F72, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32],
-    /* 18 */ [F32, ___, ___, ___, ___, ___, F72, ___, ___, ___, F72, ___, ___, F72, ___, ___, ___, F32, F32, ___, ___, ___, F72, ___, ___, ___, ___, F32],
-    /* 19 */ [F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, F32],
-    /* 20 */ [F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32],
-  ]
-
-  return visualGridToBgGridData(grid)
-}
-
-// ============================================================================
-// Export Sample BG Data Map
-// ============================================================================
-
 /**
  * Layer visualization box frame - simple border to show BG GRAPHIC boundaries
  * Used with printableArea sample to visualize all 3 screen layers
@@ -645,6 +240,10 @@ function createLayerBoxGrid(): BgGridData {
   return grid
 }
 
+// ============================================================================
+// Export Sample BG Data Map
+// ============================================================================
+
 /**
  * Map of sample keys to their pre-designed BG data
  * Only samples that use VIEW need BG data
@@ -666,33 +265,8 @@ export const SAMPLE_BG_DATA: Record<string, BgGridData> = {
   // Layer visualization - box frame for printableArea sample
   layerBox: createLayerBoxGrid(),
 
-  // ============================================================================
   // F-BASIC Reference Manual Sample Games (pages 94-101)
-  // ============================================================================
-
-  // KNIGHT - 8x8 chessboard with labels (page 94)
-  knight: createKnightGrid(),
-
-  // SUPER MEMORY - Four corner panels with center box (page 95)
-  superMemory: createSuperMemoryGrid(),
-
-  // UFO - Starry sky with mountain terrain (page 96)
-  ufo: createUfoGrid(),
-
-  // ROUTE 66 - Road with dashed center lines (page 97)
-  route66: createRoute66Grid(),
-
-  // TYPE MASTER - Text boxes and frames (page 98)
-  typeMaster: createTypeMasterGrid(),
-
-  // TURTLE - Racing track with lanes (page 99)
-  turtle: createTurtleGrid(),
-
-  // CARD - 6x6 card grid frame (page 100)
-  card: createCardGrid(),
-
-  // SCR$ Sample - Maze with flags and bricks (page 101)
-  scrSample: createScrSampleGrid(),
+  ...MANUAL_SAMPLE_BG_GRIDS,
 }
 
 /**
