+++ src/core/samples/sampleBgManualGrids.ts
@@ -1,0 +1,327 @@
+/* eslint-disable max-len */
+/**
+ * F-BASIC Reference Manual Sample Game BG Data (pages 94-101)
+ *
+ * Pre-designed BG graphics from the F-BASIC reference manual.
+ * Each function creates a 28x21 grid using visual string representations.
+ */
+
+import type { BgGridData } from '@/features/bg-editor/types'
+
+import {
+  ___,
+  // ASCII constants
+  _1_, _2_, _3_, _4_, _5_, _6_, _7_, _8_,
+  _A_, _B_, _C_, _D_, _E_, _F_, _G_, _H_, _I_, _J_, _K_, _L_, _M_, _N_, _O_, _P_, _R_, _S_, _T_, _U_, _Y_,
+  D41, D50, D52,
+  F32, F41, F61, F72,
+  G00, G10, G20, G30, G40, G42, G52, G62,
+  I60, I62, I63, I70, I72, I73,
+  J00, J02, J03, J10, J12, J13, J20, J22, J23, J30, J32, J33, J50, J70,
+K02, K12, K22, K32, K42, K50,
+K52, K62, 
+  // BG code constants
+  K72,   L02, L12, L22, L40,
+  M11, M31, M70, M71, M72,
+  visualGridToBgGridData,
+} from './sampleBgVisualGrid'
+
+// ============================================================================
+// Record type for the manual grids collection
+// ============================================================================
+
+/** Map of sample keys to their manual grid creator functions */
+export interface ManualBgGrids {
+  knight: BgGridData
+  superMemory: BgGridData
+  ufo: BgGridData
+  route66: BgGridData
+  typeMaster: BgGridData
+  turtle: BgGridData
+  card: BgGridData
+  scrSample: BgGridData
+}
+
+// ============================================================================
+// Grid Creator Functions
+// ============================================================================
+
+/**
+ * KNIGHT game BG (page 94)
+ * 8x8 chessboard with labels A-H and 8-1
+ */
+function createKnightGrid(): BgGridData {
+  const grid: string[][] = [
+    //       0    1    2    3    4    5    6    7    8    9   10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
+    /*  0 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /*  1 */ [___, _8_, ___, _7_, ___, _6_, ___, _5_, ___, _4_, ___, _3_, ___, _2_, ___, _1_, ___, ___, ___, I60, J30, J30, J30, J30, J30, J30, I70, ___],
+    /*  2 */ [K72, K52, K22, K52, K22, K52, K22, K52, K22, K52, K22, K52, K22, K52, K22, K52, L02, ___, ___, J20, _K_, _N_, _I_, _G_, _H_, _T_, J20, ___],
+    /*  3 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _A_, ___, J00, J30, J30, J30, J30, J30, J30, J10, ___],
+    /*  4 */ [K42, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /*  5 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _B_, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /*  6 */ [K42, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /*  7 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _C_, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /*  8 */ [K42, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /*  9 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _D_, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 10 */ [K42, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 11 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _E_, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 12 */ [K42, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 13 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _F_, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 14 */ [K42, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 15 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _G_, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 16 */ [K42, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K02, K52, K32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 17 */ [K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, ___, K62, _H_, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 18 */ [L12, K52, K12, K52, K12, K52, K12, K52, K12, K52, K12, K52, K12, K52, K12, K52, L22, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 19 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 20 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+  ]
+
+  return visualGridToBgGridData(grid)
+}
+
+/**
+ * SUPER MEMORY game BG (page 95)
+ * Four L-shaped corner panels pointing toward center
+ */
+function createSuperMemoryGrid(): BgGridData {
+  const grid: string[][] = [
+    //         0    1    2    3    4    5    6    7    8    9    10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
+    /*  0 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /*  1 */ [___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, I60, J30, J30, J30, J30, J30, I70],
+    /*  2 */ [___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, J20, _S_, _U_, _P_, _E_, _R_, J50],
+    /*  3 */ [___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, J20, _M_, _E_, _M_, _O_, _R_, _Y_],
+    /*  4 */ [___, K52, K52, K52, L22, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, L12, K52, K52, K52, ___, J00, J30, J30, J30, J30, J30, J70],
+    /*  5 */ [___, K52, K52, K52, K52, L22, K62, ___, ___, ___, ___, ___, ___, ___, K62, L12, K52, K52, K52, K52, ___, ___, ___, ___, ___, ___, ___, ___],
+    /*  6 */ [___, K52, K52, K52, K52, K52, L22, ___, ___, ___, ___, ___, ___, ___, L12, K52, K52, K52, K52, K52, ___, ___, ___, ___, ___, ___, ___, ___],
+    /*  7 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /*  8 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /*  9 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, K72, K52, K52, K52, L02, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 10 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, K62, ___, ___, ___, K62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 11 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, L12, K52, K52, K52, L22, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 12 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 13 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 14 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 15 */ [___, K52, K52, K52, K52, K52, L02, ___, ___, ___, ___, ___, ___, ___, K72, K52, K52, K52, K52, K52, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 16 */ [___, K52, K52, K52, K52, L02, K62, ___, ___, ___, ___, ___, ___, ___, K62, K72, K52, K52, K52, K52, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 17 */ [___, K52, K52, K52, L02, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, K72, K52, K52, K52, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 18 */ [___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 19 */ [___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 20 */ [___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, K62, K62, K62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+  ]
+
+  return visualGridToBgGridData(grid)
+}
+
+/**
+ * UFO game BG (page 96)
+ * Starry sky with mountain terrain at bottom
+ */
+function createUfoGrid(): BgGridData {
+  const grid: string[][] = [
+    //         0    1    2    3    4    5    6    7    8    9    10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
+    /*  0 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /*  1 */ [___, ___, ___, M11, M31, ___, ___, ___, ___, G62, ___, ___, G52, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /*  2 */ [___, F41, D41, D41, D41, D41, F61, ___, ___, ___, ___, ___, ___, ___, ___, G62, ___, ___, ___, G62, ___, ___, ___, ___, G52, ___, ___, ___],
+    /*  3 */ [___, ___, K12, K52, K52, K12, ___, ___, ___, ___, ___, ___, G62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /*  4 */ [___, G52, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G62, ___, ___, ___, ___, ___, G62],
+    /*  5 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /*  6 */ [___, ___, ___, G62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /*  7 */ [___, ___, ___, ___, ___, ___, G62, ___, ___, ___, G62, ___, ___, ___, ___, ___, ___, ___, G52, ___, ___, ___, ___, G62, ___, ___, ___, ___],
+    /*  8 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G62, ___, ___],
+    /*  9 */ [___, G62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 10 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G52, G62, ___, ___, ___, ___, ___, G62, ___, G52, ___, ___, ___],
+    /* 11 */ [___, ___, ___, ___, ___, ___, ___, G52, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G62, ___],
+    /* 12 */ [___, ___, ___, G62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G62, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 13 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 14 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G00, G10, ___, ___, ___, ___],
+    /* 15 */ [___, ___, ___, ___, ___, ___, ___, ___, G00, G10, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, G00, G10, G00, G20, G30, G10, ___, ___, ___],
+    /* 16 */ [___, ___, ___, ___, ___, ___, ___, G00, G20, G30, G10, ___, ___, ___, ___, ___, ___, ___, G00, G20, G20, G20, G20, G30, G30, G10, ___, ___],
+    /* 17 */ [G40, M70, M70, G40, G40, M70, G20, G20, G20, G40, G30, G40, G40, M70, G40, M70, G40, G40, G40, G40, G20, G20, G40, G40, G30, G30, G40, G40],
+    /* 18 */ [M70, G40, M70, M70, G40, G40, G40, G40, M70, M70, G40, G40, M70, G40, M70, M70, M70, G40, M70, M70, G40, G40, M70, G40, G40, G40, M70, M70],
+    /* 19 */ [F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32],
+    /* 20 */ [M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72, M72],
+  ]
+
+  return visualGridToBgGridData(grid)
+}
+
+/**
+ * ROUTE 66 game BG (page 97)
+ * Road with edges - all rows same pattern
+ */
+function createRoute66Grid(): BgGridData {
+  const grid: string[][] = [
+    //         0    1    2    3    4    5    6    7    8    9    10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
+    /*  0 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
+    /*  1 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
+    /*  2 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
+    /*  3 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
+    /*  4 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
+    /*  5 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
+    /*  6 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
+    /*  7 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
+    /*  8 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
+    /*  9 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
+    /* 10 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
+    /* 11 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
+    /* 12 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
+    /* 13 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
+    /* 14 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
+    /* 15 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
+    /* 16 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
+    /* 17 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
+    /* 18 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
+    /* 19 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
+    /* 20 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, M70, L40, M71, M71, M71, M71, M71, M71, M71, L40, M70, ___, ___, ___],
+  ]
+
+  return visualGridToBgGridData(grid)
+}
+
+/**
+ * TYPE MASTER game BG (page 98)
+ * Text boxes and frames for typing display
+ */
+function createTypeMasterGrid(): BgGridData {
+  const grid: string[][] = [
+    //         0    1    2    3    4    5    6    7    8    9    10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
+    /*  0 */ [___, K72, K52, K52, K52, K52, L02, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /*  1 */ [___, K62, _T_, _Y_, _P_, _E_, K62, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /*  2 */ [___, L12, K52, K52, K52, K52, L22, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /*  3 */ [___, I60, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, I70, ___, ___, ___, ___],
+    /*  4 */ [___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J20, ___, ___, ___, ___],
+    /*  5 */ [___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J20, ___, ___, ___, ___],
+    /*  6 */ [___, J00, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J30, J10, ___, ___, ___, ___],
+    /*  7 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /*  8 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /*  9 */ [___, I60, J30, J30, J30, J30, J30, J30, J30, J30, J30, I70, I62, J32, I72, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 10 */ [___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, J20, J22, ___, J22, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 11 */ [___, J00, J30, J30, J30, J30, J30, J30, J30, J30, J30, J10, J02, J32, J12, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 12 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 13 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 14 */ [___, I63, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, I73, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 15 */ [___, J23, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J23, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 16 */ [___, J23, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J23, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 17 */ [___, J03, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J33, J13, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 18 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 19 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 20 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+  ]
+
+  return visualGridToBgGridData(grid)
+}
+
+/**
+ * TURTLE game BG (page 99)
+ * Racing track with lanes and finish line
+ */
+function createTurtleGrid(): BgGridData {
+  const grid: string[][] = [
+    //         0    1    2    3    4    5    6    7    8    9    10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
+    /*  0 */ [D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50, D50],
+    /*  1 */ [D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52, D52],
+    /*  2 */ [G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42, G42],
+    /*  3 */ [G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52],
+    /*  4 */ [G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52, G52],
+    /*  5 */ [K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K52, K52, K52, K22],
+    /*  6 */ [___, ___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
+    /*  7 */ [_1_, ___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
+    /*  8 */ [K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K52, K52, K52, K02],
+    /*  9 */ [___, ___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
+    /* 10 */ [_2_, ___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
+    /* 11 */ [K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K52, K52, K52, K02],
+    /* 12 */ [___, ___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
+    /* 13 */ [ _3_,___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
+    /* 14 */ [K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K52, K52, K52, K02],
+    /* 15 */ [___, ___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
+    /* 16 */ [ _4_,___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
+    /* 17 */ [K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K52, K52, K52, K02],
+    /* 18 */ [___, ___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
+    /* 19 */ [ _5_,___, ___, J20, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, J22, ___, K62],
+    /* 20 */ [K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K50, K52, K52, K52, K12],
+  ]
+
+  return visualGridToBgGridData(grid)
+}
+
+/**
+ * CARD game BG (page 100)
+ * 6x6 card grid frame
+ */
+function createCardGrid(): BgGridData {
+  const grid: string[][] = [
+    //         0    1    2    3    4    5    6    7    8    9    10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
+    /*  0 */ [K72, K52, K52, L02, K72, K52, K52, L02, K72, K52, K52, L02, K72, K52, K52, L02, K72, K52, K52, L02, K72, K52, K52, L02, ___, ___, ___, ___],
+    /*  1 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
+    /*  2 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
+    /*  3 */ [K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, ___, ___, ___, ___],
+    /*  4 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
+    /*  5 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
+    /*  6 */ [K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, ___, ___, ___, ___],
+    /*  7 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
+    /*  8 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
+    /*  9 */ [K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, F72, ___, ___, ___],
+    /* 10 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
+    /* 11 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
+    /* 12 */ [K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, F72, ___, ___, ___],
+    /* 13 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
+    /* 14 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
+    /* 15 */ [K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, K42, K52, K52, K32, ___, ___, ___, ___],
+    /* 16 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
+    /* 17 */ [K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, K62, ___, ___, K62, ___, ___, ___, ___],
+    /* 18 */ [L12, K52, K52, L22, L12, K52, K52, L22, L12, K52, K52, L22, L12, K52, K52, L22, L12, K52, K52, L22, L12, K52, K52, L22, ___, ___, ___, ___],
+    /* 19 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+    /* 20 */ [___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___],
+  ]
+
+  return visualGridToBgGridData(grid)
+}
+
+/**
+ * SCR$ Sample game BG (page 101)
+ * Maze with flags (F72) and brick walls (F32)
+ */
+function createScrSampleGrid(): BgGridData {
+  const grid: string[][] = [
+    //         0    1    2    3    4    5    6    7    8    9    10   11   12   13   14   15   16   17   18   19   20   21   22   23   24   25   26   27
+    /*  0 */ [F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32],
+    /*  1 */ [F32, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32],
+    /*  2 */ [F32, ___, ___, F72, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, F72, ___, ___, ___, F72, ___, ___, ___, F72, ___, ___, ___, ___, F32],
+    /*  3 */ [F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32],
+    /*  4 */ [F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, F32, F32, F32, F32, F32, F32, F32, ___, ___, ___, ___, F32],
+    /*  5 */ [F32, ___, ___, ___, ___, ___, ___, ___, F72, ___, ___, ___, ___, ___, F32, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, F32],
+    /*  6 */ [F32, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, F72, ___, ___, F32, F32, ___, F72, ___, F32],
+    /*  7 */ [F32, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32],
+    /*  8 */ [F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32],
+    /*  9 */ [F32, ___, ___, ___, F72, ___, ___, ___, ___, ___, ___, ___, ___, F72, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, F32],
+    /* 10 */ [F32, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, F72, ___, ___, ___, ___, F32],
+    /* 11 */ [F32, F32, F32, F32, F32, F32, F32, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32],
+    /* 12 */ [F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32],
+    /* 13 */ [F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, F72, ___, ___, ___, ___, ___, ___, F72, ___, F32],
+    /* 14 */ [F32, ___, F72, ___, ___, ___, ___, ___, ___, ___, F72, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32],
+    /* 15 */ [F32, ___, ___, ___, ___, F32, F32, F32, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, F32],
+    /* 16 */ [F32, ___, ___, ___, ___, F32, F32, F32, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, F32],
+    /* 17 */ [F32, ___, F72, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32],
+    /* 18 */ [F32, ___, ___, ___, ___, ___, F72, ___, ___, ___, F72, ___, ___, F72, ___, ___, ___, F32, F32, ___, ___, ___, F72, ___, ___, ___, ___, F32],
+    /* 19 */ [F32, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, ___, F32, F32, ___, ___, ___, ___, ___, ___, ___, ___, F32],
+    /* 20 */ [F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32, F32],
+  ]
+
+  return visualGridToBgGridData(grid)
+}
+
+// ============================================================================
+// Export All Manual Grids
+// ============================================================================
+
+/** F-BASIC reference manual sample game BG data (pages 94-101) */
+export const MANUAL_SAMPLE_BG_GRIDS: ManualBgGrids = {
+  knight: createKnightGrid(),
+  superMemory: createSuperMemoryGrid(),
+  ufo: createUfoGrid(),
+  route66: createRoute66Grid(),
+  typeMaster: createTypeMasterGrid(),
+  turtle: createTurtleGrid(),
+  card: createCardGrid(),
+  scrSample: createScrSampleGrid(),
+}
